requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-openai>=0.0.1
langchain-community>=0.1.0
pytesseract>=0.3.10
pdf2image>=1.16.0
opencv-python>=4.8.0
Pillow>=10.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pydantic>=2.0.0
typing-extensions>=4.0.0
textblob>=0.17.1